Factor image renderers into separate modules, loaded on demand. For consistency, add --pil option to scripts/qr to force PIL/PNG output.
