+++ qrcode/image/pil.py
@@ -1,0 +1,34 @@
+# Try to import PIL in either of the two ways it can be installed.
+try:
+    from PIL import Image, ImageDraw
+except ImportError:
+    import Image, ImageDraw
+
+import qrcode.image.base
+
+
+class PilImage(qrcode.image.base.BaseImage):
+    """PIL image builder, default format is PNG."""
+
+    def __init__(self, border, width, box_size):
+        if Image is None and ImageDraw is None:
+            raise NotImplementedError("PIL not available")
+        super(PilImage, self).__init__(border, width, box_size)
+        self.kind = "PNG"
+
+        pixelsize = (self.width + self.border * 2) * self.box_size
+        self._img = Image.new("1", (pixelsize, pixelsize), "white")
+        self._idr = ImageDraw.Draw(self._img)
+
+    def drawrect(self, row, col):
+        x = (col + self.border) * self.box_size
+        y = (row + self.border) * self.box_size
+        box = [(x, y),
+               (x + self.box_size - 1,
+                y + self.box_size - 1)]
+        self._idr.rectangle(box, fill="black")
+
+    def save(self, stream, kind=None):
+        if kind is None:
+            kind = self.kind
+        self._img.save(stream, kind)
